server/pyproject.toml
[project]
name = "akn-profiler-server"
version = "0.1.0"
description = "Python language server for AKN Profiler — parses Akoma Ntoso XSD schemas, generates Pydantic models, and serves YAML profile editing features over LSP."
requires-python = ">=3.10"
license = { text = "SEE LICENSE IN ../LICENSE" }

dependencies = [
    "pygls>=2.0,<3.0",
    "lsprotocol>=2024.0.0",
    "pydantic>=2.6,<3.0",
    "xsdata[cli,lxml]>=24.0",
    "lxml>=5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.3", "mypy>=1.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["akn_profiler"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
